feat(Lotto): Lotto 검증 구현
