+++ README.md
@@ -16,7 +16,7 @@
 
 - [ ] 랜덤 로또 번호를 생생하는 기능
 
-- [ ] 로또 번호를 오름차순 정렬하는 기능
+- [x] 로또 번호를 오름차순 정렬하는 기능
 
 - [ ] 로또 번호를 당첨 번호와 비교하는 기능
 
@@ -34,11 +34,11 @@
 
 ### 검증
 
-- [ ] 로또 번호의 숫자 범위가 1~45인지 확인
+- [x] 로또 번호의 숫자 범위가 1~45인지 확인
 
 - [ ] 보너스 번호의 숫자 범위가 1~45인지 확인
 
-- [ ] 한 로또에 중복된 번호가 있지 않은지 확인
+- [x] 한 로또에 중복된 번호가 있지 않은지 확인
 
 - [ ] 보너스 번호가 당첨 로또와 중복되지 않은지 확인
